requirements_django.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0
django-crispy-forms>=2.0.0
crispy-bootstrap5>=0.7.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL (optional)

# Background Tasks
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Network Monitoring
ping3>=4.0.0
speedtest-cli>=2.1.0
requests>=2.31.0

# Excel Integration
openpyxl>=3.1.0
pandas>=2.0.0
xlsxwriter>=3.1.0

# Email
django-ses>=3.5.0  # AWS SES (optional)

# Utilities
python-dotenv>=1.0.0
python-decouple>=3.8.0
Pillow>=10.0.0

# Development
django-debug-toolbar>=4.1.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0

# Monitoring & Logging
sentry-sdk>=1.32.0
